address_book/requirements.txt
flask
flask-cors
mysql-connector-python
openpyxl